Finished converting first tab into pyGT
The functions for each button in the first tab have been successfully re-integrated into pyGT version of GUI.

In the previous commit, the instructions for the back light colour change was accidentally added instead of the correct instructions to add the Three point lighting. It has been resolved in this commit.